+++ settings/ContextSettings.cpp
@@ -25,6 +25,8 @@
 #include <utils/layoutbuilder.h>
 
 #include "SettingsConstants.hpp"
+#include "SettingsTr.hpp"
+#include "SettingsUtils.hpp"
 
 namespace QodeAssist::Settings {
 ContextSettings &contextSettings()
