+++ settings/SettingsUtils.hpp
@@ -29,11 +29,6 @@
 
 namespace QodeAssist::Settings {
 
-struct Tr
-{
-    Q_DECLARE_TR_FUNCTIONS(QtC::QodeAssist)
-};
-
 inline bool pingUrl(const QUrl &url, int timeout = 5000)
 {
     if (!url.isValid()) {
@@ -71,26 +66,5 @@ void resetAspect(AspectType &aspect)
 {
     aspect.setVolatileValue(aspect.defaultValue());
 }
-
-class ButtonAspect : public Utils::BaseAspect
-{
-    Q_OBJECT
-
-public:
-    ButtonAspect(Utils::AspectContainer *container = nullptr)
-        : Utils::BaseAspect(container)
-    {}
-
-    void addToLayout(Layouting::Layout &parent) override
-    {
-        auto button = new QPushButton(m_buttonText);
-        connect(button, &QPushButton::clicked, this, &ButtonAspect::clicked);
-        parent.addItem(button);
-    }
-
-    QString m_buttonText;
-signals:
-    void clicked();
-};
 
 } // namespace QodeAssist::Settings
